LPS-88355 Move classes to util component, since we want to reuse it in other test modules, not only asset test
