+++ modules/apps/asset/asset-test-util/bnd.bnd
@@ -1,4 +1,6 @@
 Bundle-Name: Liferay Asset Test Utilities
 Bundle-SymbolicName: com.liferay.asset.test.util
 Bundle-Version: 2.0.0
-Export-Package: com.liferay.asset.test.util
+Export-Package:\
+	com.liferay.asset.test.util,\
+	com.liferay.asset.test.util.asset.renderer.factory
