+++ modules/apps/asset/asset-test-util/build.gradle
@@ -4,8 +4,12 @@ dependencies {
 	compileOnly group: "com.liferay.portal", name: "com.liferay.portal.test", version: "default"
 	compileOnly group: "com.liferay.portal", name: "com.liferay.portal.test.integration", version: "default"
 	compileOnly group: "commons-lang", name: "commons-lang", version: "2.6"
+	compileOnly group: "javax.portlet", name: "portlet-api", version: "3.0.0"
 	compileOnly group: "javax.servlet", name: "javax.servlet-api", version: "3.0.1"
+	compileOnly group: "org.osgi", name: "org.osgi.core", version: "5.0.0"
+	compileOnly group: "org.osgi", name: "org.osgi.service.component.annotations", version: "1.3.0"
 	compileOnly project(":apps:asset:asset-api")
+	compileOnly project(":core:petra:petra-string")
 	compileOnly project(":core:registry-api")
 }
