+++ modules/apps/asset/asset-test-util/src/main/java/com/liferay/asset/test/util/asset/renderer/factory/TestAssetRendererModel.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.asset.assetrendererfactoryregistryutil.test;
+package com.liferay.asset.test.util.asset.renderer.factory;
 
 /**
  * @author Máté Thurzó
