+++ modules/apps/asset/asset-test-util/src/main/resources/com/liferay/asset/test/util/asset/renderer/factory/packageinfo
@@ -1,0 +1,1 @@
+version 1.0.0
